pyproject.toml
[tool.poetry]
name = "quantumguard"
version = "0.1.0-dev"
description = "Privacy-preserving, self-optimizing AI agent framework for enterprise cybersecurity"
authors = ["Declan Loder <declan@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "quantumguard", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.12.5"
structlog = "^24.1"
pyyaml = "^6.0"
pydantic = "^2.5"
networkx = "^3.2"
plotly = "^5.18"
click = "8.1.7"

[tool.poetry.scripts]
quantumguard = "quantumguard.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
ml = [
    "torch-geometric (>=2.7.0,<3.0.0)",
    "torch (==2.3.1)",
    "torchvision (==0.18.1)",
    "torchaudio (==2.3.1)",
    "flwr (>=1.26.1,<2.0.0)"
]
tools = [
    "scapy (>=2.7.0,<3.0.0)"
]
dev = [
    "streamlit (>=1.54.0,<2.0.0)"
]
